Quality of life improvement: point coordinate can now be specified as offset from the far edge of the page.
Far edge is the edge opposite to the one with zero coordinate. It is bottom for Y coordinates and right for X coordinates.
This change make it easier to reuse geometry definitions for different paper sizes as the paper width and height can only needs to be specified once now, as most geometry features are anchored to a certain paper edge.